Adapted backend change to coupon

+++ app/src/main/java/com/example/guided_project_ss21_a02_kauf_lokal_20_frontend/model/Coupon.kt
@@ -9,6 +9,6 @@ data class Coupon(
     var description: String,
     var couponCode: Int,
     var expiryDate: Date,
-    var value: Double,
+    var value: Float,
     var created: Date
 )
